+++ src/main/java/com/prgrms/coretime/timetable/domain/repository/timetable/TimetableRepository.java
@@ -1,4 +1,4 @@
-package com.prgrms.coretime.timetable.domain.repository;
+package com.prgrms.coretime.timetable.domain.repository.timetable;
 
 import com.prgrms.coretime.timetable.domain.timetable.Timetable;
 import org.springframework.data.jpa.repository.JpaRepository;
