+++ src/main/java/com/prgrms/coretime/timetable/service/TimetableService.java
@@ -4,7 +4,7 @@ package com.prgrms.coretime.timetable.service;
 import com.prgrms.coretime.common.error.NotFoundException;
 import com.prgrms.coretime.timetable.domain.Semester;
 import com.prgrms.coretime.timetable.domain.repository.TemporaryUserRepository;
-import com.prgrms.coretime.timetable.domain.repository.TimetableRepository;
+import com.prgrms.coretime.timetable.domain.repository.timetable.TimetableRepository;
 import com.prgrms.coretime.timetable.domain.timetable.Timetable;
 import com.prgrms.coretime.timetable.dto.request.TimetableCreateRequest;
 import com.prgrms.coretime.timetable.dto.request.TimetableUpdateRequest;
