+++ src/test/java/com/prgrms/coretime/timetable/domain/repository/TimetableRepositoryTest.java
@@ -4,6 +4,7 @@ import static com.prgrms.coretime.timetable.domain.Semester.FIRST;
 import static com.prgrms.coretime.timetable.domain.Semester.SECOND;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.prgrms.coretime.timetable.domain.repository.timetable.TimetableRepository;
 import com.prgrms.coretime.timetable.domain.timetable.Timetable;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
